+++ hermes/Resources/general/JinjaTransform/executer.py
@@ -1,8 +1,12 @@
 from ...executers.abstractExecuter import abstractExecuter
-import shutil
-import os, sys, stat
+import os
+import re
+import pathlib
 
-class jinjaExecuter(abstractExecuter):
+import numpy
+from jinja2 import FileSystemLoader, Environment
+
+class jinjaTransform(abstractExecuter):
 
     def _defaultParameters(self):
         return dict(
@@ -15,14 +19,29 @@ class jinjaExecuter(abstractExecuter):
 
     def _getTemplate(self,templateName,additionalTemplatePath=[]):
 
-        templatePath = [os.path.join(pathlib.Path(__file__).parent.absolute(), "jinjaTemplates")] + list(additionalTemplatePath)
-
-        tmp_path = os.path.join(pathlib.Path(__file__).parent.absolute(), "jinjaTemplates")
-        print("tmp_path = " + str (tmp_path) + "\n")
-        print("additionalTemplatePath = " + str(additionalTemplatePath) + "\n")
-        print("templatePath = " + str (templatePath) + "\n")
-        print("templateName = " + templateName + "\n")
-
+        templatePath = [pathlib.Path(__file__).parent.parent.parent.absolute()] + list(additionalTemplatePath)
         file_loader = FileSystemLoader(templatePath)
         env = Environment(loader=file_loader)
         return env.get_template(templateName)
+
+
+    def run(self, **inputs):
+        # get the  name of the template
+        templateName = inputs['template']
+        additionalTemplatePath = [os.path.abspath(x) for x in numpy.atleast_1d(inputs.get("path",[]))]
+
+        # make sure the splits are with slash
+        delimiters = ".", "/"
+        regexPattern = '|'.join(map(re.escape, delimiters))
+        spltList = re.split(regexPattern, templateName)
+        templateName = '/'.join(spltList)
+
+        # get the values to update in the template
+        values = inputs['parameters']
+
+        template = self._getTemplate(templateName,additionalTemplatePath=additionalTemplatePath)
+
+        # render jinja for the choosen template
+        output = template.render(**values)
+
+        return dict(openFOAMfile=output)
