+++ hermes/Resources/openFOAM/mesh/utils.jinja
@@ -1,0 +1,7 @@
+{%- macro toPoint(vertices) -%}
+  ({%- for coord in vertices %} {{ coord }} {%- endfor %})
+{%- endmacro %}
+
+{%- macro toBool(var) -%}
+  {%- if var%}true{%- else%}false{%endif%}
+{%- endmacro %}
